rename qat into quant in filenames only (#25194)
test=develop
