+++ paddle/fluid/inference/tests/api/analyzer_quant_image_classification_tester.cc
