+++ python/paddle/fluid/contrib/slim/tests/README.md
