+++ python/paddle/fluid/contrib/slim/tests/quant2_int8_image_classification_comparison.py
