+++ python/paddle/fluid/contrib/slim/tests/quant2_int8_nlp_comparison.py
