+++ python/paddle/fluid/contrib/slim/tests/quant_int8_image_classification_comparison.py
