+++ python/paddle/fluid/contrib/slim/tests/test_quant2_int8_mkldnn_pass.py
